backend/pyproject.toml
[tool.poetry]
name = "esclavizador-backend"
version = "0.1.0"
description = "Time tracking system backend API"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.13"
fastapi = "0.121.1"
tortoise-orm = "0.25.1"
aerich = "0.9.2"
asyncpg = "0.30.0"
pyjwt = "2.10.1"
uvicorn = {extras = ["standard"], version = "^0.34.0"}
python-multipart = "^0.0.20"
pydantic-settings = "^2.7.0"
email-validator = "^2.3.0"
pwdlib = {extras = ["argon2"], version = "^0.3.0"}

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.2"
pytest-cov = "^6.0.0"
httpx = "^0.28.1"
black = "^24.10.0"
ruff = "^0.8.4"
mypy = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py313"]
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
]
ignore = []

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "tortoise.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "aerich.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
minversion = "8.0"
addopts = "-ra -q --strict-markers"
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
filterwarnings = [
    # Ignore pytest-asyncio warnings about event loop policy (Python 3.14 compatibility issue)
    "ignore:.*asyncio.get_event_loop_policy.*:DeprecationWarning:pytest_asyncio.*",
    "ignore:.*asyncio.set_event_loop_policy.*:DeprecationWarning:pytest_asyncio.*",
    # Ignore pytest-asyncio warning about custom event_loop fixture (intentional for session-scoped DB)
    "ignore:.*event_loop fixture.*:DeprecationWarning:pytest_asyncio.*",
    # Ignore pytest-asyncio warning about unclosed event loop (we handle cleanup properly)
    "ignore:.*unclosed event loop.*:DeprecationWarning:pytest_asyncio.*",
    # Ignore event loop policy warning from our conftest.py (Python 3.14 compatibility, same as above)
    "ignore:.*asyncio.get_event_loop_policy.*:DeprecationWarning:.*conftest.*",
]

[tool.aerich]
tortoise_orm = "app.core.database.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
